some fixes in division logic

+++ division.cpp
@@ -2,10 +2,10 @@
 #include <iostream>
 
 // Function to divide two integers safely
-double divide(int a, int b) {
+int divide(int a, int b) {
     if (b == 0) {
         std::cerr << "Error: Division by zero!" << std::endl;
         return 0; // or handle error differently if needed
     }
-    return static_cast<double>(a) / b;
+    return static_cast<int>(a) / b;
 }
